refactor: 내 화면/원격 화면 컴포넌트로 미디어 로직을 이관
