+++ apps/frontend/src/feature/room/components/MyParticipantVideo.tsx
@@ -1,4 +1,9 @@
-import { ParticipantVideo, type VideoDisplayMode } from './ParticipantVideo';
+import { useRef } from 'react';
+
+import type { VideoDisplayMode } from '../types';
+import { useLocalParticipantMediaState } from '../hooks/useLocalParticipantMediaState';
+import { useVideoElementBinding } from '../hooks/useVideoElementBinding';
+import { ParticipantVideoView } from './ParticipantVideoView';
 
 interface MyParticipantVideoProps {
   id: string;
@@ -21,17 +26,38 @@ export function MyParticipantVideo({
   isAudioMuted,
   isSpeaking,
 }: MyParticipantVideoProps) {
+  const videoRef = useRef<HTMLVideoElement>(null);
+
+  const {
+    activeStream,
+    isVideoEnabled,
+    isAudioMuted: resolvedAudioMuted,
+  } = useLocalParticipantMediaState({
+    localStream: stream,
+    localCameraOn: isCameraOn,
+    localAudioMuted: isAudioMuted,
+  });
+
+  const { showOverlay } = useVideoElementBinding({
+    videoRef,
+    mode,
+    activeStream,
+    isVideoEnabled,
+  });
+
   return (
-    <ParticipantVideo
+    <ParticipantVideoView
       id={id}
       name={name}
       mode={mode}
       isCurrentUser={true}
       onModeChange={onModeChange}
-      stream={stream}
-      isCameraOn={isCameraOn}
-      isAudioMuted={isAudioMuted}
+      isAudioMuted={resolvedAudioMuted}
+      isVideoEnabled={isVideoEnabled}
+      isCurrentlyVisible={true}
       isSpeaking={isSpeaking}
+      showOverlay={showOverlay}
+      videoRef={videoRef}
     />
   );
 }
