+++ apps/frontend/src/feature/room/hooks/useRemoteParticipantMediaState.ts
@@ -1,0 +1,25 @@
+import { selectRemoteVideoStreamByParticipant, useMediaStore } from '../stores/useMediaStore';
+import { useRoomStore } from '../stores/useRoomStore';
+
+interface UseRemoteParticipantMediaStateParams {
+  id: string;
+}
+
+interface UseRemoteParticipantMediaStateResult {
+  activeStream: MediaStream | null;
+  isVideoEnabled: boolean;
+  isAudioMuted: boolean;
+}
+
+export function useRemoteParticipantMediaState({
+  id,
+}: UseRemoteParticipantMediaStateParams): UseRemoteParticipantMediaStateResult {
+  const remoteStream = useMediaStore(selectRemoteVideoStreamByParticipant(id));
+  const remoteAudioMuted = useRoomStore((state) => state.participantAudioMuted.get(id) ?? true);
+
+  return {
+    activeStream: remoteStream,
+    isVideoEnabled: !!remoteStream,
+    isAudioMuted: remoteAudioMuted,
+  };
+}
